Block Class (inheritance) and New Game Mechanic
Ambi and Laser now extend the Block class for ease in drawing and
testing bounds.  As well a new game mechanic is added called the
DeceptiBot, which moves sporadically any time a key is pressed,
attempting to make the used think that it is their character.  However,
the DeceptiBot is slightly smaller than the Ambi, and occasionally
produces “error” instead of a code.  At the end of the ams, the
DeceptiBot will reveal itself.